Add DRAM check

+++ src/perma.cpp
@@ -114,6 +114,8 @@ int main(int argc, char** argv) {
     throw e;
   }
 
-  std::filesystem::remove_all(pmem_directory / "*");
+  if (!use_dram) {
+    std::filesystem::remove_all(pmem_directory / "*");
+  }
   return 0;
 }
